fix(extension): close the v0.5.1 pre-release review findings
- Verb desync (red line): changing 手动处理方式 now syncs the bubble label
  (setVerb) and re-renders every non-pending badge, so no button reads 隐藏
  while executing 拉黑.
- Detection order: whitelist short-circuits before everything; the synced
  list outranks the v0.4-era read-only cache, so stale cached verdicts can
  no longer mask confirmed list hits or red-badge appealed accounts.
- Manual X-action failures annotate the 处理记录 row like the auto path.
- anchorByKey now registers hit accounts only (was: every scanned author —
  unbounded DOM retention on long timelines).
- Drop dead v0.4 code: heuristic() + keyword regexes in detect.ts and the
  disconnected graphql-users module (KnownUser type moved into detect.ts).
- CHANGELOG: add the 0.5.1 section (includes the autoExpand toggle from the
  sibling PR).
