+++ extension/lib/settings.ts
@@ -38,7 +38,7 @@ export interface Settings {
   categoryActions: CategoryActions; // per-category automatic action on list hits
   autoProcess: boolean; // master kill-switch for categoryActions auto hide/mute/block
   autoScope: AutoScope; // where auto actions may fire (replies-only by default)
-  edgeBase: string; // advanced: override the public-list site base URL (links only)
+  edgeBase: string; // advanced: override the service base URL — list/whitelist sync source, whitelist-apply backend AND outbound links
 }
 
 export const DEFAULT_CATEGORY_ACTIONS: CategoryActions = {
